Move SensitiveBytes to sensitive mod
